refactor: view 를 담당하는 클래스 view 패키지로 이동
